% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scouting.R
\name{dv_auto_codes}
\alias{dv_auto_codes}
\title{Add automatic codes to the scouted codes from a rally}
\usage{
dv_auto_codes(code, meta)
}
\arguments{
\item{code}{character: a character vector of scouted codes for a rally. The codes must be non-compound but need only be the first 6 characters (i.e. up to and including the evaluation_code). The code vector must have a valid point code at the end (the \verb{*p} or \code{ap} code)}

\item{meta}{list: meta component from a datavolley object}
}
\value{
A character vector of codes. This will be the same as the input \code{code} vector, but potentially with additional entries inserted before the last entry
}
\description{
Add automatic codes to the scouted codes from a rally
}
