% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv_creation.R
\name{dv_create_file_meta}
\alias{dv_create_file_meta}
\title{Create the \code{file_meta} component of a datavolley object}
\usage{
dv_create_file_meta(
  file_type = "indoor",
  date_format = "\%Y/\%m/\%d",
  generator_day,
  generator_idp,
  generator_prg,
  generator_release,
  generator_version,
  generator_name
)
}
\arguments{
\item{file_type}{string: "indoor", "beach"}

\item{date_format}{string: preferred date format}

\item{generator_day}{datetime: date and time of file creation}

\item{generator_idp, generator_prg, generator_release, generator_version, generator_name}{string: information about the software that generated the file. If missing, defaults will be used}
}
\value{
A tibble
}
\description{
Create the \code{file_meta} component of a datavolley object
}
