% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv_creation.R
\name{dv_create_meta}
\alias{dv_create_meta}
\title{Create metadata component of a datavolley object}
\usage{
dv_create_meta(
  match,
  more,
  comments,
  result,
  teams,
  players_h,
  players_v,
  video,
  attacks = dv_default_attack_combos(simplified = TRUE),
  setter_calls = dv_default_setter_calls(),
  winning_symbols = dv_default_winning_symbols()
)
}
\arguments{
\item{match}{tibble: as returned by \code{\link[=dv_create_meta_match]{dv_create_meta_match()}}}

\item{more}{tibble: as returned by \code{\link[=dv_create_meta_more]{dv_create_meta_more()}}}

\item{comments}{tibble: as returned by \code{\link[=dv_create_meta_comments]{dv_create_meta_comments()}}}

\item{result}{tibble: as returned by \code{\link[=dv_create_meta_result]{dv_create_meta_result()}}}

\item{teams}{tibble: as returned by \code{\link[=dv_create_meta_teams]{dv_create_meta_teams()}}}

\item{players_h}{tibble: home team table as returned by \code{\link[=dv_create_meta_players]{dv_create_meta_players()}}}

\item{players_v}{tibble: visiting team table as returned by \code{\link[=dv_create_meta_players]{dv_create_meta_players()}}}

\item{video}{tibble: as returned by \code{\link[=dv_create_meta_video]{dv_create_meta_video()}}}

\item{attacks}{tibble: as returned by \code{\link[=dv_create_meta_attack_combos]{dv_create_meta_attack_combos()}}}

\item{setter_calls}{tibble: as returned by \code{\link[=dv_create_meta_setter_calls]{dv_create_meta_setter_calls()}}}

\item{winning_symbols}{tibble: as returned by \code{\link[=dv_default_winning_symbols]{dv_default_winning_symbols()}}}
}
\value{
A list
}
\description{
Create metadata component of a datavolley object
}
\examples{
 dv_create_meta(match = dv_create_meta_match(zones_or_cones = "Z"), more = dv_create_meta_more(),
                comments = dv_create_meta_comments(),
                teams = dv_create_meta_teams(
                    team_ids = c("ardv", "badg"), teams = c("Aardvarks", "Badgers")))

}
