% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv_creation.R
\name{dv_create_meta_attack_combos}
\alias{dv_create_meta_attack_combos}
\title{Create the \code{attacks} metadata component of a datavolley object}
\usage{
dv_create_meta_attack_combos(
  code,
  start_zone,
  side = NA_character_,
  tempo,
  description,
  colour = NA_character_,
  start_coordinate = NA_integer_,
  target_attacker
)
}
\arguments{
\item{code}{character: vector of two-character (uppercase) attack codes. Attack codes must start with C, I, J, I, L, P, V, W, X, Y, or Z}

\item{start_zone}{integer: vector of start zones}

\item{side}{character: vector of L, C, R}

\item{tempo}{character: vector of tempo codes F, H, M, N, O, Q, T, U}

\item{description}{character: vector of attack descriptions}

\item{colour}{character: vector of colour codes (maybe) used when plotting, in "#RRGGBB" format}

\item{start_coordinate}{integer: vector of start locations as single-index coordinates (see e.g. \code{datavolley:dv_xy2index})}

\item{target_attacker}{character: vector of single-character codes giving the target attacker: B (back/right side attacker), C (centre), F (front/left side attacker), P (pipe), S (setter attack)}
}
\value{
A tibble
}
\description{
The \code{attacks} metadata component of a datavolley object describes the attack combination codes
}
