% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv_creation.R
\name{dv_create_meta_comments}
\alias{dv_create_meta_comments}
\title{Create the \code{comments} metadata component of a datavolley object}
\usage{
dv_create_meta_comments(
  summary,
  match_description,
  home_coach_comments,
  visiting_coach_comments
)
}
\arguments{
\item{summary, match_description, home_coach_comments, visiting_coach_comments}{string: comments}
}
\value{
A tibble
}
\description{
Create the \code{comments} metadata component of a datavolley object
}
