% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv_creation.R
\name{dv_create_meta_match}
\alias{dv_create_meta_match}
\title{Create the \code{match} metadata component of a datavolley object}
\usage{
dv_create_meta_match(
  date,
  season,
  league,
  phase,
  home_away,
  day_number,
  match_number,
  regulation = "indoor rally point",
  zones_or_cones
)
}
\arguments{
\item{date}{Date or datetime: match date and time}

\item{season}{string: season name}

\item{league}{string: league name}

\item{phase}{string: match phase (e.g. "Playoffs", "Finals")}

\item{home_away}{string: typically "Home" or "Away"}

\item{day_number, match_number}{numeric: day and match number}

\item{regulation}{string: "indoor sideout", "indoor rally point" (default), or "beach rally point"}

\item{zones_or_cones}{string: are attacks being scouted with "zones" or "cones" (or "Z" or "C")}
}
\value{
A tibble
}
\description{
Create the \code{match} metadata component of a datavolley object
}
