% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv_creation.R
\name{dv_create_meta_match_id}
\alias{dv_create_meta_match_id}
\title{Create the \code{match_id} metadata component of a datavolley object}
\usage{
dv_create_meta_match_id(mx)
}
\arguments{
\item{mx}{list: the \code{meta} component of a datavolley object}
}
\value{
A string
}
\description{
The \code{match_id} is generated from the \code{match} and \code{teams} components of the \code{mx} object, so ensure that these are fully populated before generating a match_id
}
