% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv_creation.R
\name{dv_create_meta_more}
\alias{dv_create_meta_more}
\title{Create the \code{more} metadata component of a datavolley object}
\usage{
dv_create_meta_more(referees, spectators, receipts, city, arena, scout)
}
\arguments{
\item{referees, spectators, receipts, city, arena}{string: match descriptors}

\item{scout}{string: name of the scout}
}
\value{
A tibble
}
\description{
Create the \code{more} metadata component of a datavolley object
}
