% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv_creation.R
\name{dv_create_meta_players}
\alias{dv_create_meta_players}
\title{Create the \code{players_h} or \code{players_v} metadata component of a datavolley object}
\usage{
dv_create_meta_players(players)
}
\arguments{
\item{players}{data.frame: the team roster, as a data.frame. \code{players} must contain the columns \code{lastname}, \code{firstname}, and \code{number}. Columns \code{player_id}, \code{role}, \code{nickname}, \code{special_role}, \code{foreign}, and \code{starting_position_set1} - \code{starting_position_set5} are optional}
}
\value{
A tibble
}
\description{
Create the \code{players_h} or \code{players_v} metadata component of a datavolley object
}
