% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv_creation.R
\name{dv_create_meta_result}
\alias{dv_create_meta_result}
\title{Create the \code{result} metadata component of a datavolley object}
\usage{
dv_create_meta_result(
  home_team_scores = NA_integer_,
  visiting_team_scores = NA_integer_,
  durations = NA_integer_
)
}
\arguments{
\item{home_team_scores}{integer: vector of home team scores in each set}

\item{visiting_team_scores}{integer: vector of visiting team scores in each set}

\item{durations}{integer: vector of set durations (in minutes). If not provided, these will be calculated from video times by \code{\link[=dv_update_meta]{dv_update_meta()}}}
}
\value{
A tibble
}
\description{
Note that the returned object has some NA values. It is expected that the user will call \code{\link[=dv_update_meta]{dv_update_meta()}} on the full \code{datavolley} object, which will update entries in this metadata component.
}
