% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv_creation.R
\name{dv_create_meta_setter_calls}
\alias{dv_create_meta_setter_calls}
\title{Create the \code{sets} (setter calls) metadata component of a datavolley object}
\usage{
dv_create_meta_setter_calls(
  code,
  description,
  colour = NA_character_,
  start_coordinate = NA_integer_,
  mid_coordinate = NA_integer_,
  end_coordinate = NA_integer_,
  path = NA_character_,
  path_colour = NA_character_
)
}
\arguments{
\item{code}{character: vector of two-character (uppercase) setter call codes. Setter call codes must start with K}

\item{description}{character: vector of setter call descriptions}

\item{colour}{character: vector of colour codes (maybe) used when plotting, in "#RRGGBB" format}

\item{start_coordinate, mid_coordinate, end_coordinate}{integer: vector of start/mid/end locations as single-index coordinates (see e.g. \code{datavolley:dv_xy2index})}

\item{path}{character: vector of paths, each a comma-separated list of single-index coordinates}

\item{path_colour}{character: vector of colour codes in "#RRGGBB" format}
}
\value{
A tibble
}
\description{
Create the \code{sets} (setter calls) metadata component of a datavolley object
}
