% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv_creation.R
\name{dv_create_meta_teams}
\alias{dv_create_meta_teams}
\alias{dv_create_meta_teams2}
\title{Create the \code{teams} metadata component of a datavolley object}
\usage{
dv_create_meta_teams(
  team_ids,
  teams,
  sets_won,
  coaches,
  assistants,
  shirt_colours
)

dv_create_meta_teams2(
  home_team_id,
  home_team,
  home_coach,
  home_assistant,
  home_shirt_colour,
  visiting_team_id,
  visiting_team,
  visiting_coach,
  visiting_assistant,
  visiting_shirt_colour,
  sets_won
)
}
\arguments{
\item{team_ids}{character: (required) 2-element vector of home and visiting team IDs}

\item{teams}{character: (required) 2-element vector of home and visiting team names}

\item{sets_won}{integer: 2-element vector of the number of sets won by the home team and visiting team}

\item{coaches}{character: 2-element vector of home and visiting team coach names}

\item{assistants}{character: 2-element vector of home and visiting team assistant coach names}

\item{shirt_colours}{character: 2-element vector of home and visiting team shirt colours in "#RRGGBB" format}

\item{home_team_id, visiting_team_id}{string: home and visiting team IDs}

\item{home_team, visiting_team}{string: home and visiting team names}

\item{home_coach, visiting_coach}{string: home and visiting coach name}

\item{home_assistant, visiting_assistant}{string: home and visiting assistant coach name}

\item{home_shirt_colour, visiting_shirt_colour}{string: home and visiting shirt colours in "#RRGGBB# format}
}
\value{
A tibble
}
\description{
\code{dv_create_meta_teams2} provides an alternative parameterization with home and visiting team information specified separately.
}
