% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv_creation.R
\name{dv_create_meta_video}
\alias{dv_create_meta_video}
\title{Create the \code{video} metadata component of a datavolley object}
\usage{
dv_create_meta_video(video_file)
}
\arguments{
\item{video_file}{character: path to video file. More than one is allowed, but anything other than the first might be ignored by some functions}
}
\value{
A data.frame
}
\description{
Create the \code{video} metadata component of a datavolley object
}
