% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scouting_decode.R
\name{dv_decode_evaluation}
\alias{dv_decode_evaluation}
\title{Decode evaluation codes}
\usage{
dv_decode_evaluation(
  skill,
  evaluation_code,
  table,
  data_type = "indoor",
  style = "default"
)
}
\arguments{
\item{skill}{character: full skill names (Serve, Reception, etc)}

\item{evaluation_code}{character: evaluation codes (#, +, etc)}

\item{table}{data.frame: optional table with columns \code{skill}, \code{evaluation_code}, \code{evaluation}. If not provided, the default table for \code{data_type} and \code{style} will be used}

\item{data_type}{string: "indoor" or "beach"}

\item{style}{string: "default", "volleymetrics"}
}
\value{
A character vector of skill types, with a "dvmessages" attribute if any issues were found
}
\description{
Decode evaluation codes
}
\examples{
dv_decode_evaluation("Serve", "#")
dv_decode_evaluation("Attack", "!")

}
