% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scouting_decode.R
\name{dv_decode_num_players}
\alias{dv_decode_num_players}
\title{Decode number of players}
\usage{
dv_decode_num_players(
  skill,
  num_players_code,
  evaluation,
  table,
  data_type = "indoor",
  style = "default"
)
}
\arguments{
\item{skill}{character: full skill names (Serve, Reception, etc)}

\item{num_players_code}{character: number of players codes (0, 1, etc)}

\item{table}{data.frame: optional table with columns \code{skill}, \code{num_players_code}, \code{num_players}. If not provided, the default table for \code{data_type} and \code{style} will be used}

\item{data_type}{string: "indoor" or "beach"}

\item{style}{string: "default", "volleymetrics"}
}
\value{
A character vector of number of players, with a "dvmessages" attribute if any issues were found
}
\description{
Decode number of players
}
\examples{
dv_decode_num_players("Reception", 7)
dv_decode_num_players("Attack", 3, data_type = "beach")

}
