% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scouting_decode.R
\name{dv_decode_skill_subtype}
\alias{dv_decode_skill_subtype}
\title{Decode skill subtypes}
\usage{
dv_decode_skill_subtype(
  skill,
  skill_subtype_code,
  evaluation,
  table,
  data_type = "indoor",
  style = "default"
)
}
\arguments{
\item{skill}{character: full skill names (Serve, Reception, etc)}

\item{skill_subtype_code}{character: subtype codes (H, P, etc)}

\item{evaluation}{character: skill evaluations (Ace, Perfect pass, etc). Only used to make small adjustments for certain scouting styles}

\item{table}{data.frame: optional table with columns \code{skill}, \code{skill_subtype_code}, \code{skill_subtype}. If not provided, the default table for \code{data_type} and \code{style} will be used}

\item{data_type}{string: "indoor" or "beach"}

\item{style}{string: "default", "volleymetrics"}
}
\value{
A character vector of skill subtypes, with a "dvmessages" attribute if any issues were found
}
\description{
Decode skill subtypes
}
\examples{
dv_decode_skill_subtype("Reception", "M")
dv_decode_skill_subtype("Attack", "P")
dv_decode_skill_subtype("Dig", "Q")

}
