% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scouting_decode.R
\name{dv_decode_skill_type}
\alias{dv_decode_skill_type}
\title{Decode skill type (tempo) codes}
\usage{
dv_decode_skill_type(
  skill,
  skill_type_code,
  table,
  data_type = "indoor",
  style = "default"
)
}
\arguments{
\item{skill}{character: full skill names (Serve, Reception, etc)}

\item{skill_type_code}{character: skill type (tempo) codes (Q, M, H, etc)}

\item{table}{data.frame: optional table with columns \code{skill}, \code{skill_type_code}, \code{skill_type}. If not provided, the default table for \code{data_type} and \code{style} will be used}

\item{data_type}{string: "indoor" or "beach"}

\item{style}{string: "default", "volleymetrics"}
}
\value{
A character vector of skill types, with a "dvmessages" attribute if any issues were found
}
\description{
Decode skill type (tempo) codes
}
\examples{
dv_decode_skill_type("Serve", "Q")
dv_decode_skill_type("Attack", "Z")

}
