% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scouting_decode.R
\name{dv_decode_special_code}
\alias{dv_decode_special_code}
\title{Decode special codes}
\usage{
dv_decode_special_code(
  skill,
  special_code,
  evaluation,
  table,
  data_type = "indoor",
  style = "default"
)
}
\arguments{
\item{skill}{character: full skill names (Serve, Reception, etc)}

\item{special_code}{character: special codes (Z, U, etc)}

\item{table}{data.frame: optional table with columns \code{skill}, \code{special_code}, \code{special}. If not provided, the default table for \code{data_type} and \code{style} will be used}

\item{data_type}{string: "indoor" or "beach"}

\item{style}{string: "default", "volleymetrics"}
}
\value{
A character vector of special code interpretations, with a "dvmessages" attribute if any issues were found
}
\description{
Decode special codes
}
\examples{
dv_decode_special_code("Reception", "P", NA_character_) ## evaluation is irrelevant here
dv_decode_special_code("Attack", "N", "Error")
dv_decode_special_code("Attack", "N", "Winning attack") ## different interpretation if not an error
dv_decode_special_code("Attack", "C", "Winning attack") ## "C" is not valid for winning attack
dv_decode_special_code("Attack", "C", "Positive, good attack") ## but is for attacks in play
dv_decode_special_code("Attack", "Q", "Positive, good attack") ## this is invalid everywhere

}
