% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scouting_defaults.R
\name{dv_default_attack_combos}
\alias{dv_default_attack_combos}
\title{Default attack combination codes table}
\usage{
dv_default_attack_combos(data_type = "indoor", simplified = TRUE)
}
\value{
A tibble
}
\description{
Default attack combination codes table
}
