% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scouting_defaults.R
\name{dv_default_num_players}
\alias{dv_default_num_players}
\title{Default number of players table}
\usage{
dv_default_num_players(data_type = "indoor", style = "default")
}
\arguments{
\item{data_type}{string: "indoor", "beach"}

\item{style}{string: conventions "default", "volleymetrics", "german"}
}
\value{
A tibble
}
\description{
Default number of players table
}
