% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scouting_defaults.R
\name{dv_default_scouting_table}
\alias{dv_default_scouting_table}
\title{Default scouting (type and evaluation for each skill) table}
\usage{
dv_default_scouting_table()
}
\value{
A tibble
}
\description{
Default scouting (type and evaluation for each skill) table
}
