% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scouting_defaults.R
\name{dv_default_skill_evaluations}
\alias{dv_default_skill_evaluations}
\title{Default skill evaluation table}
\usage{
dv_default_skill_evaluations(data_type = "indoor", style = "default")
}
\arguments{
\item{data_type}{string: "indoor", "beach"}

\item{style}{string: conventions "default", "volleymetrics", "german"}
}
\value{
A tibble
}
\description{
Default skill evaluation table
}
