% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scouting_defaults.R
\name{dv_default_skill_types}
\alias{dv_default_skill_types}
\title{Default skill_type (tempo) table}
\usage{
dv_default_skill_types(data_type = "indoor", style = "default")
}
\arguments{
\item{data_type}{string: "indoor", "beach"}

\item{style}{string: conventions "default", "volleymetrics", etc}
}
\value{
A tibble
}
\description{
Default skill_type (tempo) table
}
