% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scouting_defaults.R
\name{dv_default_winning_symbols}
\alias{dv_default_winning_symbols}
\title{Default winning symbols table}
\usage{
dv_default_winning_symbols()
}
\value{
A tibble
}
\description{
Default winning symbols table
}
