% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv_utils.R
\name{dv_get_player_meta}
\alias{dv_get_player_meta}
\title{Extract player metadata by team and jersey number}
\usage{
dv_get_player_meta(team, number, meta)
}
\arguments{
\item{team}{character: vector of "*" or "a"}

\item{number}{integer: vector of player jersey numbers}

\item{meta}{list: the \code{meta} component of a datavolley object, or a datavolley object}
}
\value{
A data frame with columns "team" and plus those in \code{meta}, sorted to match the inputs \code{team} and \code{number}
}
\description{
Extract player metadata by team and jersey number
}
