% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_skills.R
\name{dv_insert_sets_check}
\alias{dv_insert_sets_check}
\title{Find attacks for which we could insert missing setting actions}
\usage{
dv_insert_sets_check(
  x,
  no_set_attacks = c("PP", "P2", "PR"),
  phase_select = "Reception"
)
}
\arguments{
\item{x}{datavolley: datavolley object as returned by \code{\link[datavolley:read_dv]{datavolley::dv_read()}}}

\item{no_set_attacks}{character: vector of attack codes for which we will not automatically insert sets (e.g. setter tips, overpass attacks)}

\item{phase_select}{character: play phase(s) of attacks to consider. One or more of "Reception", "Transition"}
}
\value{
The row numbers of attacks in the \code{plays} component of x, for which sets could be inserted
}
\description{
Find attacks for which we could insert missing setting actions
}
