% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scouting.R
\name{dv_update_meta}
\alias{dv_update_meta}
\title{Update the metadata component of a datavolley object}
\usage{
dv_update_meta(x)
}
\arguments{
\item{x}{datavolley: as returned by e.g. \code{datavolley::dv_read()}}
}
\value{
A modified copy of \code{x}
}
\description{
These elements will be updated:
\itemize{
\item the \code{played}, \code{duration}, and score-related columns in \code{x$meta$result}
\item the \code{sets_won} and \code{won_match} columns in \code{x$meta$teams}
\item the starting positions and substitutions in \code{x$meta$players_h} and \code{x$meta$players_v}
}
}
