% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perana_defaults.R
\name{pv_default_errortypes}
\alias{pv_default_errortypes}
\title{Define errortype interpretations}
\usage{
pv_default_errortypes(data_type)
}
\arguments{
\item{data_type}{string: "indoor" or "beach"}
}
\value{
a tibble with columns "skill", "errortype", and "evaluation"
}
\description{
Define the interpretation of errortypes associated with events.
}
\examples{
pv_default_errortypes("beach")

}
