% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perana_defaults.R
\name{pv_default_eventgrades}
\alias{pv_default_eventgrades}
\title{Define eventgrade interpretations}
\usage{
pv_default_eventgrades(data_type)
}
\arguments{
\item{data_type}{string: "indoor" or "beach"}
}
\value{
a tibble with columns "skill", "eventgrade", "evaluation_code" (the equivalent DataVolley code, if there is one), "evaluation", and "win_loss"
}
\description{
Define the interpretation of eventgrades associated with events.
}
\examples{
pv_default_eventgrades("beach")

}
