% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perana_defaults.R
\name{pv_default_eventtypes}
\alias{pv_default_eventtypes}
\title{Lookup tables for Perana Sports database codes}
\usage{
pv_default_eventtypes(data_type)
}
\arguments{
\item{data_type}{string: "indoor" or "beach"}
}
\value{
A tibble
}
\description{
Lookup tables for Perana Sports database codes
}
