% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perana_defaults.R
\name{pv_default_subevents}
\alias{pv_default_subevents}
\title{Define subevent interpretations}
\usage{
pv_default_subevents(data_type)
}
\arguments{
\item{data_type}{string: "indoor" or "beach"}
}
\value{
a tibble with columns "skill", "subevent", and "interpretation"
}
\description{
Define the interpretation of subevents associated with events
}
\examples{
pv_default_subevents("beach")

}
