% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perana_defaults.R
\name{pv_default_subevents2}
\alias{pv_default_subevents2}
\title{Define subevent2 interpretations}
\usage{
pv_default_subevents2(data_type)
}
\arguments{
\item{data_type}{string: "indoor" or "beach"}
}
\value{
a tibble with columns "skill", "subevent2", and "interpretation"
}
\description{
Define the interpretation of subevents2 associated with events
}
\examples{
pv_default_subevents("beach")

}
