% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{pv_parse_ballstring}
\alias{pv_parse_ballstring}
\title{Parse a ballstring found in Perana Sports files}
\usage{
pv_parse_ballstring(z, which = "start")
}
\arguments{
\item{z}{character: vector of ballstring entries}

\item{which}{string: "start", "mid", or "end". Will be used to name the columns of the returned data.frame}
}
\value{
A data.frame with columns \code{start_coordinate_x} and \code{start_coordinate_y} (or \verb{mid_*} or \verb{end_*}, depending on \code{which})
}
\description{
Parse a ballstring found in Perana Sports files
}
\examples{
pv_parse_ballstring(c("51.25, 143.33"))

}
