% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scouting_defaults.R
\name{dv_default_attack_combos}
\alias{dv_default_attack_combos}
\title{Default attack combination codes table}
\usage{
dv_default_attack_combos(
  data_type = "indoor",
  style = "default",
  simplified = TRUE
)
}
\arguments{
\item{data_type}{string: "indoor", "beach"}

\item{style}{string: conventions "default", "volleymetrics", "german"}

\item{simplified}{logical: if \code{TRUE}, just the most common ones (indoor only)}
}
\value{
A tibble
}
\description{
Default attack combination codes table
}
