% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scouting_defaults.R
\name{dv_default_setter_calls}
\alias{dv_default_setter_calls}
\title{Default setter calls table}
\usage{
dv_default_setter_calls(data_type = "indoor")
}
\arguments{
\item{data_type}{string: "indoor", "beach"}
}
\value{
A tibble
}
\description{
Default setter calls table
}
