% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scouting_defaults.R
\name{dv_default_winning_symbols}
\alias{dv_default_winning_symbols}
\title{Default winning symbols table}
\usage{
dv_default_winning_symbols(data_type = "indoor", style = "default")
}
\arguments{
\item{data_type}{string: "indoor", "beach"}

\item{style}{string: conventions "default", "volleymetrics", "german"}
}
\value{
A tibble
}
\description{
Default winning symbols table
}
