% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scouting.R
\name{dv_expand_rally_codes}
\alias{dv_expand_rally_codes}
\title{Expand scouted rally codes}
\usage{
dv_expand_rally_codes(
  rx,
  last_home_setter_position,
  last_visiting_setter_position,
  last_home_team_score,
  last_visiting_team_score,
  keepcols,
  meta,
  rebuild_codes = TRUE,
  do_warn = TRUE
)
}
\arguments{
\item{rx}{data.frame: with at least the columns:
\itemize{
\item point_id, code, team ("*" or "a"), point, substitution, timeout, home_setter_position, visiting_setter_position
\item home_team_score, visiting_team_score (scores at the end of the rally)
\item player_out, player_in (substitions)
and optional columns:
\item player_number, skill (one-character code), skill_type_code, evaluation_code (used to rebuild the codes before calling \code{\link[=dv_green_codes]{dv_green_codes()}}
}}

\item{last_home_setter_position, last_visiting_setter_position}{integer: home and visiting setter positions in the previous rally}

\item{last_home_team_score, last_visiting_team_score}{integer: home and visiting team scores at the end of the previous rally}

\item{keepcols}{character: names of the columns in \code{rx} to keep, when inserting new rows. Values in these columns will be copied from an adjacent row. If \code{keepcols} is not provided, a guess will be made}

\item{meta}{list: meta component from a datavolley object}

\item{rebuild_codes}{logical: if \code{TRUE}, reconstruct the scout code before calling \code{\link[=dv_green_codes]{dv_green_codes()}}}

\item{do_warn}{logical: if \code{FALSE}, suppress warnings issued directly by this function (but warnings generated by downstream code will still occur)}
}
\value{
\code{rx} potentially with additional rows inserted. Note that the added rows might have fractional \code{point_id} values, in which case you will need to renumber all \code{point_id}s in the match to integers afterwards
}
\description{
\itemize{
\item insert dummy rally rows for score corrections. Note that this will not adjust the setter position in the dummy rows nor insert intermediate setter positions (but will insert one setter position code per team at the start of the returned data frame, if needed)
\item for a substitution, insert the setter assignment code (*PXX or aPXX) if it was the on-court setter that was substituted
\item insert the setter position codes at the start of the rally. These should not be inserted on the first point of a set, in that situation they should be in the >LUp codes - so for the first point, make sure that the \code{last_home_setter_position} and \code{last_visiting_setter_position} are passed as their starting values)
}
}
