% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filetype.R
\name{dv_file_type}
\alias{dv_file_type}
\title{Detect scout file type}
\usage{
dv_file_type(filename, error_on_unknown = FALSE)
}
\arguments{
\item{filename}{character: paths to files. Files that do not exist will be classified as "unknown" type}

\item{error_on_unknown}{logical: if \code{TRUE} an error will be thrown if an unknown file type is encountered}
}
\value{
Character vector containing:
\itemize{
\item "dvw"
\item "vsm"
\item "psvb"
\item "hxml"
\item "unknown"
}
}
\description{
Returns the general file type (format), but does not differentiate e.g. indoor from beach. Type is based on the file extension and a minimal check of the contents, but without fully parsing the file.
}
\examples{
\dontrun{
  library(ovdata)
  dv_file_type(ovdata_example("190301_kats_beds")) ## dvw
  dv_file_type(ovdata_example("clickscout")) ## dvw
  dv_file_type(ovdata_example("2017_AVL_mens_HEAT_vs_UTSSU")) ## psvb
  dv_file_type(tempfile(fileext = ".dvw")) ## unknown (contents not as expected for dvw)
}
}
